rename slidescore_api/logging.py to slidescore_api/slidescore_logging.py
